Atsiskaitymo diena

+++ CA_Final_Persons_Reg_Sys/Controllers/UsersController.cs
@@ -23,17 +23,12 @@ namespace CA_Final_Persons_Reg_Sys.Controllers
         private readonly IUserMapper _userMapper;
         private readonly IJwtService _jwtService;
 
-        //private readonly IUserRepository _userRepository;   //remove
-        //private readonly IPictureRepository _pictureRepository; //remove
-
-        public UsersController(IUserService userService, IPictureService pictureService, IUserMapper userMapper, IJwtService jwtService)//, IUserRepository userRepository, IPictureRepository pictureRepository)
+        public UsersController(IUserService userService, IPictureService pictureService, IUserMapper userMapper, IJwtService jwtService)
         {
             _userService = userService;
             _pictureService = pictureService;
             _userMapper = userMapper;
             _jwtService = jwtService;
-            //_userRepository = userRepository;
-            //_pictureRepository = pictureRepository;
         }
 
         [Authorize(Roles = "admin")]
@@ -63,6 +58,8 @@ namespace CA_Final_Persons_Reg_Sys.Controllers
         [HttpPost]
         public async Task<ActionResult<long>> CreateUser([FromForm] string requestStr, FileUploadRequest picture)  //[FromForm] FileUploadRequest picture, 
         {
+            if (string.IsNullOrEmpty(requestStr))
+                return BadRequest();
             var request = JsonConvert.DeserializeObject<UserCreateRequest>(requestStr); //NuGet package: Newtonsoft.Json
             if (request == null)
                 return BadRequest("User request is null");
@@ -85,6 +82,8 @@ namespace CA_Final_Persons_Reg_Sys.Controllers
         [HttpGet("{id}")]
         public async Task<ActionResult<UserResult>> GetUserById(long id)
         {
+            if(id <= 0)
+                return BadRequest();
             var existingUser = await _userService.GetById(id);
 
             if (existingUser == null)
@@ -98,6 +97,9 @@ namespace CA_Final_Persons_Reg_Sys.Controllers
         [HttpGet("{id}/Picture")]
         public async Task<IActionResult> GetPicture([FromRoute] long id)
         {
+            if (id <= 0)
+                return BadRequest();
+
             var pictureUrl = await _userService.GetUserPictureUrlByUserId(id);
             if (pictureUrl == null)
                 return NotFound("User or picture not found");
@@ -117,6 +119,9 @@ namespace CA_Final_Persons_Reg_Sys.Controllers
         [HttpPut("{id}/Picture")]
         public async Task<IActionResult> UpdateUserPicture([FromRoute] long id, [FromForm] FileUploadRequest request)
         {
+            if (id <= 0)
+                return BadRequest();
+
             if (request == null)
                 return BadRequest("No file uploaded.");
 
@@ -138,6 +143,9 @@ namespace CA_Final_Persons_Reg_Sys.Controllers
         [HttpPut("{id}/Password")]
         public async Task<IActionResult> UpdateUserPassword([FromRoute] long id, [FromBody] string newPassword)
         {
+            if (id <= 0)
+                return BadRequest();
+
             var existinUser = await _userService.GetById(id);
             if (existinUser == null)
                 return NotFound("User not found");
@@ -152,7 +160,8 @@ namespace CA_Final_Persons_Reg_Sys.Controllers
         [HttpPut("{id}/Name")]
         public async Task<IActionResult> UpdateUserName([FromRoute] long id, [FromBody] string newName)
         {
-            //var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (id <= 0)
+                return BadRequest();
 
             var existinUser = await _userService.GetById(id);
             if (existinUser == null)
@@ -168,6 +177,9 @@ namespace CA_Final_Persons_Reg_Sys.Controllers
         [HttpPut("{id}/LastName")]
         public async Task<IActionResult> UpdateUserLastName([FromRoute] long id, [FromBody] string newLastName)
         {
+            if (id <= 0)
+                return BadRequest();
+
             var existinUser = await _userService.GetById(id);
             if (existinUser == null)
                 return NotFound("User not found");
@@ -181,6 +193,9 @@ namespace CA_Final_Persons_Reg_Sys.Controllers
         [HttpPut("{id}/PersonalCode")]
         public async Task<IActionResult> UpdateUserPersonalCode([FromRoute] long id, [FromBody] string newPersonalCode)
         {
+            if (id <= 0)
+                return BadRequest();
+
             var existinUser = await _userService.GetById(id);
             if (existinUser == null)
                 return NotFound("User not found");
@@ -194,6 +209,9 @@ namespace CA_Final_Persons_Reg_Sys.Controllers
         [HttpPut("{id}/PhoneNumber")]
         public async Task<IActionResult> UpdateUserPhoneNumber([FromRoute] long id, [FromBody] string newPhoneNumber)
         {
+            if (id <= 0)
+                return BadRequest();
+
             var existinUser = await _userService.GetById(id);
             if (existinUser == null)
                 return NotFound("User not found");
@@ -207,6 +225,9 @@ namespace CA_Final_Persons_Reg_Sys.Controllers
         [HttpPut("{id}/Email")]
         public async Task<IActionResult> UpdateUserEmail([FromRoute] long id, [FromBody] string newEmail)
         {
+            if (id <= 0)
+                return BadRequest();
+
             var existinUser = await _userService.GetById(id);
             if (existinUser == null)
                 return NotFound("User not found");
@@ -221,6 +242,9 @@ namespace CA_Final_Persons_Reg_Sys.Controllers
         [HttpPut("{id}/CityName")]
         public async Task<IActionResult> UpdateUserCityName([FromRoute] long id, [FromBody] string newCityName)
         {
+            if (id <= 0)
+                return BadRequest();
+
             var existinUser = await _userService.GetById(id);
             if (existinUser == null)
                 return NotFound("User not found");
@@ -230,50 +254,59 @@ namespace CA_Final_Persons_Reg_Sys.Controllers
             return NoContent();
         }
 
-        //[Authorize(Roles = "user,admin")]
-        //[HttpPut("{id}/StreetName")]
-        //public async Task<IActionResult> UpdateUserStreetName([FromRoute] long id, [FromBody] string newStreetName)
-        //{
-        //    var existinUser = await _userRepository.GetByIdAsync(id);
-        //    if (existinUser == null)
-        //        return NotFound("User not found");
-        //    if (!await _userRepository.UpdateUserStreetNameAsync(id, newStreetName))
-        //        return StatusCode(500, "Failed to update street name");
+        [Authorize(Roles = "user,admin")]
+        [HttpPut("{id}/StreetName")]
+        public async Task<IActionResult> UpdateUserStreetName([FromRoute] long id, [FromBody] string newStreetName)
+        {
+            if (id <= 0)
+                return BadRequest();
+            var existinUser = await _userService.GetById(id);
+            if (existinUser == null)
+                return NotFound("User not found");
+            if (!await _userService.UpdateUserStreetName(id, newStreetName))
+                return StatusCode(500, "Failed to update street name");
 
-        //    return NoContent();
-        //}
+            return NoContent();
+        }
 
-        //[Authorize(Roles = "user,admin")]
-        //[HttpPut("{id}/HouseNumber")]
-        //public async Task<IActionResult> UpdateUserHouseNumber([FromRoute] long id, [FromBody] string newHouseNumber)
-        //{
-        //    var existinUser = await _userRepository.GetByIdAsync(id);
-        //    if (existinUser == null)
-        //        return NotFound("User not found");
-        //    if (!await _userRepository.UpdateUserHouseNumberAsync(id, newHouseNumber))
-        //        return StatusCode(500, "Failed to update house number");
+        [Authorize(Roles = "user,admin")]
+        [HttpPut("{id}/HouseNumber")]
+        public async Task<IActionResult> UpdateUserHouseNumber([FromRoute] long id, [FromBody] string newHouseNumber)
+        {
+            if (id <= 0)
+                return BadRequest();
 
-        //    return NoContent();
-        //}
+            var existinUser = await _userService.GetById(id);
+            if (existinUser == null)
+                return NotFound("User not found");
+            if (!await _userService.UpdateUserHouseNumber(id, newHouseNumber))
+                return StatusCode(500, "Failed to update house number");
 
-        //[Authorize(Roles = "user,admin")]
-        //[HttpPut("{id}/ApartmentNumber")]
-        //public async Task<IActionResult> UpdateUserApartmentNumber([FromRoute] long id, [FromBody] string newApartmentNumber)
-        //{
-        //    var existinUser = await _userRepository.GetByIdAsync(id);
-        //    if (existinUser == null)
-        //        return NotFound("User not found");
-        //    if (!await _userRepository.UpdateUserApartmentNumberAsync(id, newApartmentNumber))
-        //        return StatusCode(500, "Failed to update apartment number");
+            return NoContent();
+        }
 
-        //    return NoContent();
-        //}
+        [Authorize(Roles = "user,admin")]
+        [HttpPut("{id}/ApartmentNumber")]
+        public async Task<IActionResult> UpdateUserApartmentNumber([FromRoute] long id, [FromBody] string newApartmentNumber)
+        {
+            if (id <= 0)
+                return BadRequest();
+
+            var existinUser = await _userService.GetById(id);
+            if (existinUser == null)
+                return NotFound("User not found");
+            if (!await _userService.UpdateUserApartmentNumber(id, newApartmentNumber))
+                return StatusCode(500, "Failed to update apartment number");
+
+            return NoContent();
+        }
 
         [Authorize(Roles = "admin")]
         [HttpDelete("{id}")]
         public IActionResult DeleteUser(long id)
         {
-            //_fileService.DeleteFile(id);
+            if (id <= 0)
+                return BadRequest();
             _userService.DeleteUser(id);
             
             return NoContent();
